requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu

